+++ src/main/java/duke/Duke.java
@@ -1,5 +1,4 @@
-import java.time.LocalDate;
-import java.time.format.DateTimeParseException;
+package duke;
 
 /**
  * Represents a Personal Assistant Chatbot that helps a person keeps track of various things.
